Add files via upload
There are a multitude of applications for association rules and in this case, we used it to find the relationship between a store's products via their transactions. The data that we will receive from this dataset is about how likely a customer is to purchase two unlikely items together. All sorts of relationship and conclusions can be drawn from this such as better physical layouts, price increases, or bundle deals.